migrate all path globals to overcooked_ai_py.static
