+++ src/overcooked_ai_py/static.py
@@ -7,3 +7,4 @@ PLANNERS_DIR = os.path.join(DATA_DIR, "planners")
 LAYOUTS_DIR = os.path.join(DATA_DIR, "layouts")
 GRAPHICS_DIR = os.path.join(DATA_DIR, "graphics")
 FONTS_DIR = os.path.join(DATA_DIR, "fonts")
+TESTING_DATA_DIR = os.path.join(DATA_DIR, "testing")
